+++ src/main.rs
@@ -4,10 +4,14 @@ mod models;
 mod repository;
 mod task;
 mod utils;
+mod conn;
 
 use anyhow::Result;
 use futures::future::join_all;
-use std::sync::Arc;
+use models::CmdResultJSON;
+use std::{collections::HashMap, sync::{Arc, Mutex}};
+
+type JsonMapMutex = Mutex<HashMap<String, CmdResultJSON>>;
 
 // TODO(magurotuna): ここの値も要検討
 const JOB_THREADS: usize = 3;
@@ -18,12 +22,16 @@ async fn main() -> Result<()> {
     let config = config::load_config()?;
     let db_conn = Arc::new(db::new_pool(&config).await?);
     let docker_conn = Arc::new(bollard::Docker::connect_with_unix_defaults()?);
+    let json_map_mutex = Arc::new(Mutex::new(HashMap::new()));
+    tokio::spawn(conn::server(json_map_mutex.clone()));
+    
 
     let mut handles = Vec::new();
     for _ in 0..JOB_THREADS {
         let db = Arc::clone(&db_conn);
         let docker = Arc::clone(&docker_conn);
-        let handle = tokio::spawn(task::gen_job(db, docker));
+        let json_map = Arc::clone(&json_map_mutex);
+        let handle = tokio::spawn(task::gen_job(db, docker, json_map));
         handles.push(handle);
     }
